+++ src/soundRecorder.js
@@ -5,6 +5,7 @@ define(function (require) {
   // inspiration: recorder.js, Tone.js & typedarray.org
 
   var p5sound = require('master');
+  var convertToWav = require('helpers').convertToWav;
   var ac = p5sound.audiocontext;
 
   /**
@@ -244,101 +245,22 @@ define(function (require) {
     this._jsNode = null;
   };
 
-  /**
-   *  Export a usable url for blob object.
-   *
-   *  @method saveSoundToBlob
-   *  @param  {p5.SoundFile} soundFile p5.SoundFile that you wish to export
-   */
-  p5.prototype.saveSoundToBlob = function(soundFile) {
-    const dataView = convertToWav(soundFile)
-    const audioBlob = new Blob([dataView], {type: 'audio/wav'})
-    return URL.createObjectURL(audioBlob)
-  }
 
   /**
-   *  Save a p5.SoundFile as a .wav audio file.
+   * Save a p5.SoundFile as a .wav file. The browser will prompt the user
+   * to download the file to their device.
+   * For uploading audio to a server, use
+   * <a href="/docs/reference/#/p5.SoundFile/saveBlob">`p5.SoundFile.saveBlob`</a>.
    *
+   *  @for p5
    *  @method saveSound
    *  @param  {p5.SoundFile} soundFile p5.SoundFile that you wish to save
-   *  @param  {String} name      name of the resulting .wav file.
+   *  @param  {String} fileName      name of the resulting .wav file.
    */
-  p5.prototype.saveSound = function(soundFile, name) {
-    const dataView = convertToWav(soundFile)
-    p5.prototype.writeFile( [ dataView ], name, 'wav');
+  // add to p5.prototype as this is used by the p5 `save()` method.
+  p5.prototype.saveSound = function (soundFile, fileName) {
+    const dataView = convertToWav(soundFile.buffer);
+    p5.prototype.writeFile([dataView], fileName, 'wav');
   };
-
-  // helper methods to convert audio file as .wav format,
-  // will use as saving .wav file and saving blob object
-  function convertToWav(soundFile){
-    var leftChannel, rightChannel;
-    leftChannel = soundFile.buffer.getChannelData(0);
-
-    // handle mono files
-    if (soundFile.buffer.numberOfChannels > 1) {
-      rightChannel = soundFile.buffer.getChannelData(1);
-    } else {
-      rightChannel = leftChannel;
-    }
-
-    var interleaved = interleave(leftChannel,rightChannel);
-
-    // create the buffer and view to create the .WAV file
-    var buffer = new window.ArrayBuffer(44 + interleaved.length * 2);
-    var view = new window.DataView(buffer);
-
-    // write the WAV container,
-    // check spec at: https://ccrma.stanford.edu/courses/422/projects/WaveFormat/
-    // RIFF chunk descriptor
-    writeUTFBytes(view, 0, 'RIFF');
-    view.setUint32(4, 36 + interleaved.length * 2, true);
-    writeUTFBytes(view, 8, 'WAVE');
-    // FMT sub-chunk
-    writeUTFBytes(view, 12, 'fmt ');
-    view.setUint32(16, 16, true);
-    view.setUint16(20, 1, true);
-    // stereo (2 channels)
-    view.setUint16(22, 2, true);
-    view.setUint32(24, 44100, true);
-    view.setUint32(28, 44100 * 4, true);
-    view.setUint16(32, 4, true);
-    view.setUint16(34, 16, true);
-    // data sub-chunk
-    writeUTFBytes(view, 36, 'data');
-    view.setUint32(40, interleaved.length * 2, true);
-
-    // write the PCM samples
-    var lng = interleaved.length;
-    var index = 44;
-    var volume = 1;
-    for (var i = 0; i < lng; i++) {
-      view.setInt16(index, interleaved[i] * (0x7FFF * volume), true);
-      index += 2;
-    }
-
-    return view
-  }
-
-  // helper methods to save waves
-  function interleave(leftChannel, rightChannel) {
-    var length = leftChannel.length + rightChannel.length;
-    var result = new Float32Array(length);
-
-    var inputIndex = 0;
-
-    for (var index = 0; index < length; ) {
-      result[index++] = leftChannel[inputIndex];
-      result[index++] = rightChannel[inputIndex];
-      inputIndex++;
-    }
-    return result;
-  }
-
-  function writeUTFBytes(view, offset, string) {
-    var lng = string.length;
-    for (var i = 0; i < lng; i++) {
-      view.setUint8(offset + i, string.charCodeAt(i));
-    }
-  }
 
 });
